requirements-dev.txt
pydevd-odoo==1.2.1